+++ models/__init__.py
@@ -3,4 +3,5 @@
 from . import customs_stage
 from . import customs_document_type
 from . import customs_operation
+from . import customs_operation_line
 from . import customs_document_requirement